requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.3.5
aiohttp==3.10.2
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
asttokens==2.4.1
astunparse==1.6.3
attrs==24.2.0
certifi==2024.7.4
charset-normalizer==3.3.2
colorama==0.4.6
comm==0.2.2
dataclasses-json==0.6.7
debugpy==1.8.1
decorator==5.1.1
distro==1.9.0
django-widget-tweaks==1.5.0
executing==2.0.1
flatbuffers==24.3.25
frozenlist==1.4.1
gast==0.6.0
google-pasta==0.2.0
greenlet==3.0.3
grpcio==1.65.4
h11==0.14.0
h5py==3.11.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
ipykernel==6.29.4
ipython==8.23.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.5.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jupyter_client==8.6.1
jupyter_core==5.7.2
keras==3.4.1
langchain==0.2.12
langchain-community==0.2.11
langchain-core==0.2.29
langchain-openai==0.1.20
langchain-text-splitters==0.2.2
langsmith==0.1.98
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
matplotlib-inline==0.1.6
mdurl==0.1.2
ml-dtypes==0.4.0
multidict==6.0.5
mypy-extensions==1.0.0
mysqlclient==2.2.4
namex==0.0.8
nest-asyncio==1.6.0
numpy==1.26.4
openai==1.40.2
opt-einsum==3.3.0
optree==0.12.1
orjson==3.10.7
packaging==24.0
pandas==2.2.2
parso==0.8.4
pillow==10.4.0
platformdirs==4.2.0
prompt-toolkit==3.0.43
protobuf==4.25.4
psutil==5.9.8
pure-eval==0.2.2
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.17.2
pypdf==4.3.1
python-dateutil==2.9.0.post0
pytz==2024.1
pywin32==306
PyYAML==6.0.2
pyzmq==25.1.2
regex==2024.7.24
requests==2.32.3
rich==13.7.1
scikit-learn==1.5.1
scipy==1.14.0
setuptools==69.5.1
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.32
stack-data==0.6.3
tenacity==8.5.0
tensorboard==2.17.0
tensorboard-data-server==0.7.2
tensorflow==2.17.0
tensorflow-intel==2.17.0
termcolor==2.4.0
threadpoolctl==3.5.0
tiktoken==0.7.0
tornado==6.4
tqdm==4.66.5
traitlets==5.14.2
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.2.2
wcwidth==0.2.13
Werkzeug==3.0.3
wheel==0.43.0
wrapt==1.16.0
yarl==1.9.4
